+++ paper_uploads/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.5.2"
+__version__ = "0.6.0"
 default_app_config = "paper_uploads.apps.Config"