+++ paper_uploads/forms/widgets/image.py
@@ -1,4 +1,6 @@
 from django import forms
+from django.contrib.admin import site
+from django.contrib.admin.sites import NotRegistered
 from django.urls import reverse_lazy
 
 from .base import FileWidgetBase
@@ -24,13 +26,19 @@ class ImageWidget(FileUploaderWidgetMixin, FileWidgetBase):
     def get_context(self, name, value, attrs):
         context = super().get_context(name, value, attrs)
 
-        # urls
-        info = self.model._meta.app_label, self.model._meta.model_name
-        context.update(
-            {
-                "upload_url": reverse_lazy("admin:%s_%s_upload" % info),
-                "change_url": reverse_lazy("admin:%s_%s_change" % info),
-                "delete_url": reverse_lazy("admin:%s_%s_delete" % info),
-            }
-        )
+        # support proxy models
+        for model_class in {self.model, self.model._meta.concrete_model}:
+            if site.is_registered(model_class):
+                info = model_class._meta.app_label, model_class._meta.model_name
+                context.update(
+                    {
+                        "upload_url": reverse_lazy("admin:%s_%s_upload" % info),
+                        "change_url": reverse_lazy("admin:%s_%s_change" % info),
+                        "delete_url": reverse_lazy("admin:%s_%s_delete" % info),
+                    }
+                )
+                break
+        else:
+            raise NotRegistered("The model %s is not registered." % self.model.__name__)
+
         return context