+++ paper_uploads/models/collection.py
@@ -9,7 +9,7 @@ from django.contrib.staticfiles.finders import find
 from django.contrib.staticfiles.storage import staticfiles_storage
 from django.core import checks
 from django.core.files import File
-from django.db import models
+from django.db import models, transaction
 from django.db.models import functions
 from django.db.models.base import ModelBase
 from django.db.models.fields.files import FieldFile
@@ -24,11 +24,17 @@ from ..conf import FILE_ICON_DEFAULT, FILE_ICON_OVERRIDES, settings
 from ..helpers import _get_item_types, _set_item_types, build_variations
 from ..storage import upload_storage
 from ..variations import PaperVariation
-from .base import NoPermissionsMetaBase, ResourceBaseMeta, FileFieldResource, VersatileImageResourceMixin
+from .base import (
+    FileFieldResource,
+    NoPermissionsMetaBase,
+    ResourceBaseMeta,
+    VersatileImageResourceMixin,
+)
 from .fields import CollectionItem
 from .fields.collection import ContentItemRelation
 from .image import VariationalFileField
 from .mixins import BacklinkModelMixin
+from .utils import generate_filename
 
 __all__ = [
     "CollectionItemBase",
@@ -129,6 +135,11 @@ class CollectionBase(BacklinkModelMixin, metaclass=CollectionMeta):
         default_permissions = ()
         verbose_name = _("collection")
         verbose_name_plural = _("collections")
+
+    def delete(self, using=None, keep_parents=False):
+        # удаляем элементы вручную из-за рекурсии (см. clean_uploads.py, строка 145)
+        self.items.all().delete()
+        super().delete(using=using, keep_parents=keep_parents)
 
     @classmethod
     def _check_fields(cls, **kwargs):
@@ -400,7 +411,7 @@ class FileItem(FilePreviewMixin, CollectionFileItemBase):
         _("file"),
         max_length=255,
         storage=upload_storage,
-        upload_to=settings.COLLECTION_FILES_UPLOAD_TO,
+        upload_to=generate_filename,
     )
     display_name = models.CharField(_("display name"), max_length=255, blank=True)
 
@@ -412,6 +423,9 @@ class FileItem(FilePreviewMixin, CollectionFileItemBase):
         if not self.pk and not self.display_name:
             self.display_name = self.basename
         super().save(*args, **kwargs)
+
+    def get_file_folder(self) -> str:
+        return settings.COLLECTION_FILES_UPLOAD_TO
 
     def get_file(self) -> FieldFile:
         return self.file
@@ -436,7 +450,7 @@ class SVGItem(CollectionFileItemBase):
         _("file"),
         max_length=255,
         storage=upload_storage,
-        upload_to=settings.COLLECTION_FILES_UPLOAD_TO,
+        upload_to=generate_filename,
     )
     display_name = models.CharField(_("display name"), max_length=255, blank=True)
 
@@ -448,6 +462,9 @@ class SVGItem(CollectionFileItemBase):
         if not self.pk and not self.display_name:
             self.display_name = self.basename
         super().save(*args, **kwargs)
+
+    def get_file_folder(self) -> str:
+        return settings.COLLECTION_FILES_UPLOAD_TO
 
     def get_file(self) -> FieldFile:
         return self.file
@@ -474,12 +491,15 @@ class ImageItem(VersatileImageResourceMixin, CollectionFileItemBase):
         _("file"),
         max_length=255,
         storage=upload_storage,
-        upload_to=settings.COLLECTION_IMAGES_UPLOAD_TO,
+        upload_to=generate_filename,
     )
 
     class Meta(CollectionItemBase.Meta):
         verbose_name = _("Image item")
         verbose_name_plural = _("Image items")
+
+    def get_file_folder(self) -> str:
+        return settings.COLLECTION_IMAGES_UPLOAD_TO
 
     def get_file(self) -> FieldFile:
         return self.file