+++ tests/tests/models/test_base.py
@@ -13,9 +13,9 @@ from app.models import (
     DummyImageFieldResource,
     DummyResource,
     DummyVersatileImageResource,
-    VariationFile,
 )
 from paper_uploads import signals
+from paper_uploads.files import VariationFile
 from paper_uploads.variations import PaperVariation
 
 from .. import utils
@@ -580,6 +580,9 @@ class TestFileFieldResource(TestFileResource):
         file_name = storage.resource.name
         pattern = posixpath.join(self.resource_location, 'Nature_Tree{suffix}.Jpeg')
         assert file_name == utils.get_target_filepath(pattern, file_name)
+
+    def test_get_file_folder(self, storage):
+        assert storage.resource.get_file_folder() == ""
 
     def test_get_file_field(self, storage):
         assert (